Comment out gruser and grdate calls (for now)
The libgiza based version of libpgplot included with Ubuntu 24.04 does
not seem to provide the `gruser` and `grdate` functions, resulting in
"undefined reference" errors at link time.  These are only used to
generate an annotation string in the `plotpeak` function.  For now these
calls (and the associated `pgtext` call that uses the results) have been
commented out.  If needed, implementations of these functions could be
borrowed/copied from PGPLOT.